POLIO-2072: move polio filter out of iaso
