+++ iaso/api/common.py
@@ -23,7 +23,6 @@ from rest_framework.viewsets import ModelViewSet as BaseModelViewSet, ViewSet
 from rest_framework_csv.renderers import CSVRenderer
 
 from hat.api_import.models import APIImport
-from iaso.models import OrgUnit, OrgUnitType
 from iaso.models.payments import PaymentStatuses
 
 
@@ -416,26 +415,6 @@ class CSVExportMixin:
             content=renderer.render(data),
         )
         return response
-
-
-class CustomFilterBackend(filters.BaseFilterBackend):
-    def filter_queryset(self, request, queryset, view):
-        search = request.query_params.get("search")
-        if search:
-            country_types = OrgUnitType.objects.countries().only("id")
-            org_units = OrgUnit.objects.filter(
-                name__icontains=search, org_unit_type__in=country_types, path__isnull=False
-            ).only("id")
-
-            query = Q(obr_name__icontains=search) | Q(epid__icontains=search)
-            if len(org_units) > 0:
-                query.add(
-                    Q(initial_org_unit__path__descendants=OrgUnit.objects.query_for_related_org_units(org_units)), Q.OR
-                )
-
-            return queryset.filter(query)
-
-        return queryset
 
 
 class IsAdminOrSuperUser(permissions.BasePermission):
