+++ plugins/polio/api/campaigns/filters/search.py
@@ -1,0 +1,24 @@
+from django.db.models import Q
+from rest_framework import filters
+
+from iaso.models import OrgUnit, OrgUnitType
+
+
+class SearchFilterBackend(filters.BaseFilterBackend):
+    def filter_queryset(self, request, queryset, view):
+        search = request.query_params.get("search")
+        if search:
+            country_types = OrgUnitType.objects.countries().only("id")
+            org_units = OrgUnit.objects.filter(
+                name__icontains=search, org_unit_type__in=country_types, path__isnull=False
+            ).only("id")
+
+            query = Q(obr_name__icontains=search) | Q(epid__icontains=search)
+            if len(org_units) > 0:
+                query.add(
+                    Q(initial_org_unit__path__descendants=OrgUnit.objects.query_for_related_org_units(org_units)), Q.OR
+                )
+
+            return queryset.filter(query)
+
+        return queryset
